Add tree simplifier package
The purpose of this package is to take the parse tree generated by the
parser (`src/parse`) and convert it to a format which is easy for the
interpreter to execute (an Abstract Syntax Tree or AST). This package
also trims unnecessary parse tree nodes in the process, resulting in an
overall smaller footprint at execution time.